Subject: feat:Admin password for update and delete
Body:
Adding middleware to check a admin password for things like update and delete
for both categories and items. This is done in router before the post request
goes through.
Adding the JS script to fire the prompt and add that to the hidden input to
pass that as a request to express.
A few style changes to buttons and forms and errors for forms to display
in red to show it's need fixing.